Converted LogBot from JDA 2.2.1 to JDA 3.0
