+++ build.gradle
@@ -1,16 +1,16 @@
 apply plugin: 'java'
 
 repositories {
-    jcenter()
+       jcenter()
 }
 
-version = "1.2.2"
+version = "0.0.1"
 //create a single Jar with all dependencies (just change the MainClass)
-task fatJar(type: Jar) {
+task fat(type: Jar) {
     manifest {
         attributes 'Implementation-Title': 'LogBot',
                 'Implementation-Version': version,
-                'Main-Class': 'com.xelitexirish.logbot.LogBot'
+                'Main-Class': 'kingdgrizzle.logbot.LogBot'
     }
     baseName = project.name + '-all'
     from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
@@ -19,12 +19,11 @@ task fatJar(type: Jar) {
 
 dependencies {
 
-    compile 'net.dv8tion:JDA:2.2.1_375'
+    compile 'net.dv8tion:JDA:3.0.BETA_98'
 
     // https://mvnrepository.com/artifact/org.json/json
-    compile group: 'org.json', name: 'json', version: '20160212'
+	compile group: 'org.json', name: 'json', version: '20160810'
 
-    // https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple
-    compile group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1'
-
+	// https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple
+	compile group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
 }
