+++ settings.gradle
@@ -1,19 +1,2 @@
-/*
- * This settings file was auto generated by the Gradle buildInit task
- * by 'Shaun' at '10/10/16 22:41' with Gradle 2.14
- *
- * The settings file is used to specify which projects to include in your build.
- * In a single project build this file can be empty or even removed.
- *
- * Detailed information about configuring a multi-project build in Gradle can be found
- * in the user guide at https://docs.gradle.org/2.14/userguide/multi_project_builds.html
- */
-
-/*
-// To declare projects as part of a multi-project build use the 'include' method
-include 'shared'
-include 'api'
-include 'services:webservice'
-*/
 
 rootProject.name = 'LogBot'
