--- src/main/java/com/xelitexirish/logbot/utils/Constants.java
@@ -1,13 +1,0 @@
-package com.xelitexirish.logbot.utils;
-
-/**
- * Created by XeliteXirish on 12/10/2016. www.xelitexirish.com
- */
-public class Constants {
-
-    public static final String COMMAND_PREFIX = "/log ";
-    public static final String BOT_VERSION = "1.2.2";
-
-    public static final String XELITEXIRISH_ID = "182210823630880768";
-
-}
