+++ src/main/java/kingdgrizzle/logbot/utils/Constants.java
@@ -1,0 +1,10 @@
+package kingdgrizzle.logbot.utils;
+
+public class Constants {
+	
+	public static final String COMMAND_PREFIX = "/log ";
+    public static final String BOT_VERSION = "1.0.0";
+
+    public static final String XELITEXIRISH_ID = "182210823630880768";
+    public static final String KINGDGRIZZLE_ID = "139836912335716352";
+}
