+++ src/main/java/kingdgrizzle/logbot/utils/GeneralUtils.java
@@ -1,16 +1,13 @@
-package com.xelitexirish.logbot.utils;
+package kingdgrizzle.logbot.utils;
 
 import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 
-/**
- * Created by XeliteXirish on 23/10/2016. www.xelitexirish.com
- */
 public class GeneralUtils {
 
-    public static String getCurrentTime() {
+	public static String getCurrentTime() {
         return new SimpleDateFormat("yyyy:MM:dd - HH:mm:ss").format(Calendar.getInstance().getTime());
     }
