+++ src/main/java/kingdgrizzle/logbot/utils/MessageUtils.java
@@ -1,14 +1,12 @@
-package com.xelitexirish.logbot.utils;
+package kingdgrizzle.logbot.utils;
 
-import net.dv8tion.jda.MessageBuilder;
-import net.dv8tion.jda.Permission;
-import net.dv8tion.jda.entities.Message;
+import net.dv8tion.jda.core.MessageBuilder;
+import net.dv8tion.jda.core.Permission;
+import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.User;
 
-/**
- * Created by XeliteXirish on 12/10/2016. www.xelitexirish.com
- */
 public class MessageUtils {
-
+	
     public static Message getNoPermissionMsg(Permission permission) {
         return MessageUtils.wrapStringInCodeBlock("Sorry but you don't have the required permission to use this command. Required permission: " + permission.name());
     }
@@ -20,8 +18,14 @@ public class MessageUtils {
     }
 
     public static Message wrapStringInCodeBlock(String message, String language) {
-        MessageBuilder messageBuilder = new MessageBuilder();
+       MessageBuilder messageBuilder = new MessageBuilder();
         messageBuilder.appendCodeBlock(message, language);
         return messageBuilder.build();
+    }
+    
+    public static Message usefulError(User user, String message) {
+    	MessageBuilder messageBuilder = new MessageBuilder();
+    	messageBuilder.appendMention(user).appendString(", you haven\'t used the correct syntax. Please use: \n").appendCodeBlock(message, "");
+    	return messageBuilder.build();
     }
 }
